usb: dwc3: dwc3-msm: Check for valid EXTCON connection source
During DWC3 MSM probe, if extcon and usb role switch are both undefined,
the mode will be set to peripheral.  However, in case EUD is defined, then
the extcon parameter will be non-null, and the probe routine will avoid
setting the default mode into peripheral, even though there is no valid
source for notifying cable events.

In case of when EUD is registered as the EXTCON device, continue to allow
setting the default mode as peripheral mode, as it does not send cable
events to DWC3 MSM.  This is useful bringup scenarios where UCSI role
switch is not ready, but EUD is enabled.

Change-Id: I82263a125f67560558236356ddc733301c586de3
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Faisal Hassan <faisal.hassan@oss.qualcomm.com>